auth/pyproject.toml
[tool.poetry]
name = "georecord_auth"
version = "0.1.0"
description = "The auth API microservice"
authors = ["Thomas Day <thomas.day@uqconnect.edu.au>"]
readme = "README.md"
packages = [{include = "georecord_auth"}]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
asyncpg = "^0.29.0"
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
